method reservation added

+++ internal/entity/reservation.go
@@ -1,8 +1,8 @@
 package entity
 
 type UserReservation struct {
-	Id        int     `json:"id"`
-	IdService int     `json:"id_service"`
-	IdOrder   int     `json:"id_order"`
-	Cost      float64 `json:"cost"`
+	Id        uint32  `json:"id"`
+	IdService uint32  `json:"id_service"`
+	IdOrder   uint32  `json:"id_order"`
+	Cost      float32 `json:"cost"`
 }
